+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.0] - 2021-04-26
+
+### Changed
+  - Installation removed from README.
+  - Testing tree generation for HTML and LaTeX.
+  
 ## [2.3.0] - 2021-04-20
 
 ### Changed